requirements.txt
streamlit
plotly
pandas
feedparser
newspaper4k
groq
sentence-transformers
scikit-learn